requirements.txt
spacy
transformers
torch
sentencepiece
pandas
tqdm
regex
ftfy
streamlit 
opencv-python 
ultralytics 
deepface 
pytesseract 
Pillow 
openai
